--- spring-cloud-netflix-hystrix-amqp/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.netflix.hystrix.amqp.HystrixStreamAutoConfiguration
