+++ spring-cloud-netflix-hystrix-stream/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.springframework.cloud.netflix.hystrix.stream.HystrixStreamAutoConfiguration
