--- spring-cloud-netflix-turbine-amqp/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.netflix.turbine.amqp.TurbineAmqpAutoConfiguration
-
-org.springframework.context.ApplicationListener=\
-org.springframework.cloud.netflix.turbine.amqp.TurbinePortApplicationListener
