--- spring-cloud-netflix-turbine-amqp/src/main/resources/turbine.yml
@@ -1,14 +1,0 @@
-info:
-  component: Turbine AMQP
-spring:
-  application:
-    name: turbine
-  jmx:
-    default_domain: cloud.turbine.amqp
-
-server:
-  port: 8990
-turbine:
-  amqp:
-    port: 8989
-
