+++ spring-cloud-netflix-turbine-stream/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,5 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.springframework.cloud.netflix.turbine.stream.TurbineStreamAutoConfiguration
+
+org.springframework.context.ApplicationListener=\
+org.springframework.cloud.netflix.turbine.stream.TurbinePortApplicationListener
