+++ spring-cloud-netflix-turbine-stream/src/main/resources/turbine.yml
@@ -1,0 +1,16 @@
+info:
+  component: Turbine Stream
+spring:
+  application:
+    name: turbine
+  cloud.stream.bindings:
+    turbineStreamInput: springCloudHystrixStream
+  jmx:
+    default_domain: cloud.turbine.stream
+
+server:
+  port: 8990
+turbine:
+  stream:
+    port: 8989
+
