+++ spring-cloud-netflix-turbine-stream/src/test/resources/org/springframework/cloud/netflix/turbine/stream/hystrixamqp.stream
