+++ spring-cloud-starter-turbine-stream/src/main/resources/META-INF/spring.provides
